Add a 'remove' command to delete configs from the cache.
The CLI code was also refactored into a sub-package.

ZEN-34657
