+++ Products/ZenCollector/configcache/app/init.py
@@ -1,11 +1,15 @@
 ##############################################################################
 #
-# Copyright (C) Zenoss, Inc. 2023, all rights reserved.
+# Copyright (C) Zenoss, Inc. 2024, all rights reserved.
 #
 # This content is made available according to terms specified in
 # License.zenoss under the directory where your Zenoss product is installed.
 #
 ##############################################################################
+
+import sys
+
+import six
 
 from zope.configuration import xmlconfig
 
@@ -41,6 +45,8 @@ def _no_zope(configs, overrides):
 
 def _load_configs(ctx, configs):
     for filename, module in configs:
+        if isinstance(module, six.string_types):
+            module = sys.modules[module]
         xmlconfig.file(filename, package=module, context=ctx)
