+++ Products/ZenCollector/configcache/cli/list.zcml
