+++ Products/ZenCollector/configcache/cli/remove.zcml
