+++ Products/ZenCollector/configcache/cli/show.zcml
@@ -1,0 +1,13 @@
+<?xml version="1.0"?>
+<!--
+Copyright (C) Zenoss, Inc. 2024 all rights reserved.
+
+This content is made available according to terms specified in
+License.zenoss under the directory where your Zenoss product is installed.
+-->
+<configure xmlns="http://namespaces.zope.org/zope">
+
+   <include file="meta.zcml" package="zope.component" />
+   <include file="store.zcml" package="Products.ZenCollector.configcache" />
+
+</configure>
