+++ Products/ZenCollector/configcache/configcache.py
@@ -10,7 +10,7 @@
 from __future__ import absolute_import, print_function
 
 from .app.args import get_arg_parser
-from .cli import List_, Show, Expire
+from .cli import Expire, List_, Remove, Show
 from .invalidator import Invalidator
 from .manager import Manager
 from .version import Version
@@ -27,6 +27,7 @@ def main(argv=None):
     List_.add_arguments(parser, subparsers)
     Show.add_arguments(parser, subparsers)
     Expire.add_arguments(parser, subparsers)
+    Remove.add_arguments(parser, subparsers)
 
     args = parser.parse_args()
     args.factory(args).run()
